Feature/salt pepper (#7)
* added jwt

* working jwt

* imports shit

---------

+++ backend/server/backend.py
@@ -1,11 +1,15 @@
-from typing import Optional
+from json import load
+from pathlib import Path
 from http import HTTPStatus
 from logging import getLogger
+from typing import Optional, Union
+from bcrypt import gensalt, hashpw
 from bson import json_util, ObjectId
 
 from flask_cors import CORS
 from pymongo.errors import PyMongoError
 from flask import Flask, jsonify, request, Response
+from jwt import encode as jwt_encode, decode as jwt_decode, InvalidTokenError, ExpiredSignatureError
 
 from mongo_manager import MongoDBContextManager
 
@@ -17,10 +21,21 @@ class BackendRestServer:
     PORT = 1360
     CORS_ORIGINS = '*'
 
+    JWT_ALGORITHM = 'HS256'
+
     DATA_INDEX = 0
     STATUS_INDEX = 1
 
     def __init__(self) -> None:
+        current_dir: Path = Path(__file__).parent
+        with open(current_dir.joinpath('config.json')) as config_file:
+            config = load(config_file)
+
+        self.salt = gensalt()
+        self.pepper = config['PEPPER']
+        self.jwt_secret = config['JWT_SECRET']
+        self.admin_password = config['ADMIN_PASSWORD']
+
         self.app: Flask = Flask(__name__)
         self.setup_routes()
 
@@ -40,6 +55,9 @@ class BackendRestServer:
         self.app.add_url_rule('/applicants/<string:applicant_type>',
                               view_func=self.add_applicant,
                               methods=['POST'])
+        self.app.add_url_rule('/authenticate',
+                              view_func=self.authentication_check,
+                              methods=['GET'])
         self.app.add_url_rule('/applicants/<string:applicant_type>/<string:object_id>',
                               view_func=self.update_applicant,
                               methods=['PUT'])
@@ -62,8 +80,47 @@ class BackendRestServer:
         object_id: ObjectId = json_util.loads(serialized_object_id)
         return {'_id': object_id}
 
+    def verify_jwt(self, encoded_jwt: str) -> Union[Exception, bool]:
+        try:
+            # Decode and verify the JWT
+            jwt_decode(encoded_jwt, self.jwt_secret, algorithms=[BackendRestServer.JWT_ALGORITHM])
+            return True
+        except ExpiredSignatureError as exception:
+            server_log.exception('JWT has expired', exception)
+            return exception
+        except InvalidTokenError as exception:
+            server_log.exception('Invalid JWT', exception)
+            return exception
+
+    def create_password_hash(self, password: str) -> tuple[str, str]:
+        combined_string = self.pepper + password
+        
+        hashed_password = hashpw(combined_string.encode('utf-8'), self.salt)
+
+        return hashed_password
+
+    def verify_password(self, password: str, stored_hash: str) -> bool:
+        combined_string = self.pepper + password
+        hashed_password = hashpw(combined_string.encode('utf-8'), self.salt)
+
+        return hashed_password == stored_hash
+
+    def authentication_check(self) -> tuple[Response, HTTPStatus]:
+        if self.verify_password(request.json['password'],
+                                self.create_password_hash(self.admin_password)):
+            payload = {'user': 'admin'}
+            encoded_jwt = jwt_encode(payload, self.jwt_secret, algorithm='HS256')
+            
+            return jsonify({'token': encoded_jwt}), HTTPStatus.OK
+
+        return jsonify({'Exception': 'Incorrect password'}), HTTPStatus.FORBIDDEN
+
     def get_all_applicants(self,
                            applicant_area: Optional[str] = None) -> tuple[Response, HTTPStatus]:
+        jwt_valid = self.verify_jwt(request.headers['token'])
+        if not isinstance(jwt_valid, bool):
+            return jsonify({'Exception': jwt_valid}), HTTPStatus.FORBIDDEN
+
         with MongoDBContextManager('requests') as mongo:
             area_filter = self.create_area_filter(applicant_area)
             requests_data = list(mongo.collection.find(area_filter))
@@ -79,6 +136,10 @@ class BackendRestServer:
     def get_applicants(self,
                        applicant_type: str,
                        applicant_area: Optional[str] = None) -> tuple[Response, HTTPStatus]:
+        jwt_valid = self.verify_jwt(request.headers['Token'])
+        if not isinstance(jwt_valid, bool):
+            return jsonify({'Exception': jwt_valid}), HTTPStatus.FORBIDDEN
+
         with MongoDBContextManager(applicant_type) as mongo:
             data_list = {'data_list': list(mongo.collection.find(self.create_area_filter(applicant_area)))}
 
@@ -101,6 +162,10 @@ class BackendRestServer:
     def update_applicant(self,
                          applicant_type: str,
                          serialized_object_id: str) -> tuple[Response, HTTPStatus]:
+        jwt_valid = self.verify_jwt(request.headers['token'])
+        if not isinstance(jwt_valid, bool):
+            return jsonify({'Exception': jwt_valid}), HTTPStatus.FORBIDDEN
+
         with MongoDBContextManager(applicant_type) as mongo:
             try:
                 mongo.collection.find_one_and_replace(self.create_object_filter(serialized_object_id), request.json)
@@ -113,6 +178,10 @@ class BackendRestServer:
     def delete_applicant(self,
                          applicant_type: str,
                          object_id: str) -> tuple[Response, HTTPStatus]:
+        jwt_valid = self.verify_jwt(request.headers['token'])
+        if not isinstance(jwt_valid, bool):
+            return jsonify({'Exception': jwt_valid}), HTTPStatus.FORBIDDEN
+
         with MongoDBContextManager(applicant_type) as mongo:
             try:
                 mongo.collection.delete_one(self.create_object_filter(object_id))
@@ -125,6 +194,10 @@ class BackendRestServer:
     def patch_applicant(self,
                         applicant_type: str,
                         serialized_object_id: str) -> tuple[Response, HTTPStatus]:
+        jwt_valid = self.verify_jwt(request.headers['token'])
+        if not isinstance(jwt_valid, bool):
+            return jsonify({'Exception': jwt_valid}), HTTPStatus.FORBIDDEN
+
         with MongoDBContextManager(applicant_type) as mongo:
             try:
                 mongo.collection.find_one_and_update(self.create_object_filter(serialized_object_id), request.json)
